+++ CHANGELOG.md
@@ -3,15 +3,17 @@ Polyverse Boost Automatic Application Modernization
 
 # Release Notes
 
-## Version 1.2.4: August 2nd, 2023
+## Version 1.2.4: August 3rd, 2023
 
 ### New Features
 - Added support for analyzing Razor files (CSHTML extension, HTML with embedded C# or Visual Basic.NET)
+- Added support for analyzing Salesforce Apex files (Visual Studio Code APEX extension not required for analysis)
 
 ### Enhancements
 - Enable right-click Menus for Compiled Python files (*.pod)
 - Rebuild/reparse all source files during analysis if not analysis output will be lost
 - Errors are treated as Warnings by default in Diagnostic Problems to avoid blocking build
+- Report Canceled account status for users who have canceled their subscriptions
 
 ### Bug Fixes
 - Fix links in Summary View